backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
python-decouple>=3.8
Pillow>=10.1.0
PyPDF2>=3.0.1
pdfplumber>=0.10.3
scikit-learn>=1.3.2
pandas>=2.1.3
numpy>=1.26.2
spacy>=3.7.2
nltk>=3.8.1
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
sentry-sdk>=1.39.0
